requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-environ==0.4.5
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
whitenoise==5.1.0
reportlab==3.5.53
svglib==1.0.1
stripe==2.48.0
twilio==6.46.0
vcrpy==4.1.1
pdf2image==1.14.0
boto3==1.16.14
django-storages==1.10.1
faker==4.14.2
factory-boy==3.1.0
django-positions==0.6.0
django-ses==1.0.3
pyTelegramBotAPI==3.7.6
django-enumfield==2.0.2
python-magic==0.4.22
flake8==3.8.4
